Move symbols into Transform and minor refactor
Move the ImputeProperty and ImMethod types.

This is the first commit in the re-factoring that has involved
a code change, in this case the introduction of the imputeField_
helper to avoid having to export a number of helper symbols
from Transform to Core.
