+++ hvega/src/Graphics/Vega/VegaLite/Transform.hs
@@ -24,12 +24,16 @@ module Graphics.Vega.VegaLite.Transform
        , WOperation(..)
        , BinProperty(..)
        , WindowProperty(..)
+       , ImputeProperty(..)
+       , ImMethod(..)
 
          -- not for extgernal export
        , aggregate_
        , op_
-       , bin
        , binned_
+       , impute_
+       , imputeFields_
+       , bin
        , binProperty
        , windowFieldProperty
        , windowPropertySpec
@@ -43,7 +47,12 @@ import Data.Aeson ((.=), object, toJSON)
 import Data.Maybe (mapMaybe)
 
 
-import Graphics.Vega.VegaLite.Specification (VLSpec, LabelledSpec)
+import Graphics.Vega.VegaLite.Data
+  ( DataValue
+  , DataValues
+  , dataValueSpec
+  , dataValuesSpecs
+  )
 import Graphics.Vega.VegaLite.Foundation
   ( SortField
   , sortFieldSpec
@@ -51,6 +60,7 @@ import Graphics.Vega.VegaLite.Foundation
   , fromT
   , allowNull
   )
+import Graphics.Vega.VegaLite.Specification (VLSpec, LabelledSpec)
 
 
 {-|
@@ -365,3 +375,119 @@ windowPropertySpec wps =
       fromSpecs _ = A.Null
 
   in map fromSpecs [frms, ips, gps, sts]
+
+
+-- | This is used with 'Graphics.Vega.VegaLite.impute' and 'Graphics.Vega.VegaLite.PImpute'.
+--
+--   @since 0.4.0.0
+
+data ImputeProperty
+    = ImFrame (Maybe Int) (Maybe Int)
+      -- ^ 1d window over which data imputation values are generated. The two
+      --   parameters should either be @Just@ a number indicating the offset from the current
+      --   data object, or @Nothing@ to indicate unbounded rows preceding or following the
+      --   current data object.
+    | ImKeyVals DataValues
+      -- ^ Key values to be considered for imputation.
+    | ImKeyValSequence Double Double Double
+      -- ^ Key values to be considered for imputation as a sequence of numbers between
+      --   a start (first parameter), to less than an end (second parameter) in steps of
+      --   the third parameter.
+    | ImMethod ImMethod
+      -- ^ How is the imputed value constructed.
+      --
+      --   When using @ImMethod 'ImValue'@, the replacement value is
+      --   set with 'ImNewValue'.
+    | ImGroupBy [T.Text]
+      -- ^ Allow imputing of missing values on a per-group basis. For use with the impute
+      --   transform only and not a channel encoding.
+    | ImNewValue DataValue
+      -- ^ The replacement value (when using @ImMethod 'ImValue'@).
+
+
+imputeProperty :: ImputeProperty -> LabelledSpec
+imputeProperty (ImFrame m1 m2) = "frame" .= map allowNull [m1, m2]
+imputeProperty (ImKeyVals dVals) = "keyvals" .= dataValuesSpecs dVals
+imputeProperty (ImKeyValSequence start stop step) =
+  "keyvals" .= object ["start" .= start, "stop" .= stop, "step" .= step]
+imputeProperty (ImMethod method) = "method" .= imMethodLabel method
+imputeProperty (ImNewValue dVal) = "value" .= dataValueSpec dVal
+imputeProperty (ImGroupBy _) = "groupby" .= A.Null
+
+
+imputePropertySpecFrame, imputePropertySpecKeyVals,
+  imputePropertySpecKeyValSequence, imputePropertySpecGroupBy,
+  imputePropertySpecMethod, imputePropertySpecValue :: ImputeProperty -> Maybe VLSpec
+
+imputePropertySpecFrame (ImFrame m1 m2) = Just (toJSON (map allowNull [m1, m2]))
+imputePropertySpecFrame _ = Nothing
+
+imputePropertySpecKeyVals (ImKeyVals dVals) = Just (toJSON (dataValuesSpecs dVals))
+imputePropertySpecKeyVals _ = Nothing
+
+imputePropertySpecKeyValSequence (ImKeyValSequence start stop step) =
+  let obj = ["start" .= start, "stop" .= stop, "step" .= step]
+  in Just (object obj)
+imputePropertySpecKeyValSequence _ = Nothing
+
+imputePropertySpecGroupBy (ImGroupBy fields) = Just (toJSON fields)
+imputePropertySpecGroupBy _ = Nothing
+
+imputePropertySpecMethod (ImMethod method) = Just (toJSON (imMethodLabel method))
+imputePropertySpecMethod _ = Nothing
+
+imputePropertySpecValue (ImNewValue dVal) = Just (dataValueSpec dVal)
+imputePropertySpecValue _ = Nothing
+
+
+impute_ :: [ImputeProperty] -> LabelledSpec
+impute_ ips = "impute" .= object (map imputeProperty ips)
+
+
+imputeFields_ ::
+  T.Text
+  -- ^ The data field to process.
+  -> T.Text
+  -- ^ The key field to uniquely identify data objects within a group.
+  -> [ImputeProperty]
+  -- ^ Define how the imputation works.
+  -> LabelledSpec
+imputeFields_ fields key imProps =
+  let ags = [ fromT fields
+            , fromT key
+            , toSpec (mapMaybe imputePropertySpecFrame imProps)
+            , toSpec (mapMaybe imputePropertySpecKeyVals imProps)
+            , toSpec (mapMaybe imputePropertySpecKeyValSequence imProps)
+            , toSpec (mapMaybe imputePropertySpecMethod imProps)
+            , toSpec (mapMaybe imputePropertySpecGroupBy imProps)
+            , toSpec (mapMaybe imputePropertySpecValue imProps) ]
+
+      toSpec [x] = x
+      toSpec _ = A.Null
+
+  in "impute" .= toJSON ags
+
+
+-- | Imputation method to use when replacing values.
+--
+--   @since 0.4.0.0
+
+data ImMethod
+  = ImMin
+    -- ^ Use the minimum value.
+  | ImMax
+    -- ^ Use the maximum value.
+  | ImMean
+    -- ^ Use the mean value.
+  | ImMedian
+    -- ^ Use the median value.
+  | ImValue
+    -- ^ Use a replacement value (set with @ImNewValue@).
+
+
+imMethodLabel :: ImMethod -> T.Text
+imMethodLabel ImMin = "min"
+imMethodLabel ImMax = "max"
+imMethodLabel ImMean = "mean"
+imMethodLabel ImMedian = "median"
+imMethodLabel ImValue = "value"
